+++ CHANGELOG.md
@@ -54,6 +54,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Structure of parsed entities, to reflect documents saved in the database (#20)
 - Improved and simplified code (#22)
 - Refactored VEP annotations parsing module (#38)
+- Refactor and simplify main function to accept and parse case config file (#)
 ### Fixed
 - Normalize chromosome names by stripping the chr prefix (#4)
 - Parsing of end chromosome (#29)